% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.clustal.alignment.R
\name{read.clustal.alignment}
\alias{read.clustal.alignment}
\title{Read a clustal (.aln) alignment file}
\usage{
read.clustal.alignment(file)
}
\arguments{
\item{file}{The name of the file to be read. The file name 
should be provided within quotes, either as an absolute 
file path, or a relative file path given the current 
working directory, getwd().}
}
\value{
Returns a data frame with taxon names as row names 
     and a column for each site in the alignment.
}
\description{
This function reads a clustal (.aln) alignment file and 
     converts it to an easy-to-read data frame object.
}
\examples{
read.clustal.alignment()
}
\keyword{DNA}
\keyword{alignment}
\keyword{aln}
\keyword{clustal}
\keyword{read}
