% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.alignment.R
\name{read.fasta.alignment}
\alias{read.fasta.alignment}
\title{Read a fasta alignment file}
\usage{
read.fasta.alignment(file)
}
\arguments{
\item{file}{The name of the file to be read. The file name 
should be provided within quotes, either as an absolute 
file path, or a relative file path given the current 
working directory, getwd().}
}
\value{
Returns a data frame with taxon names as row names 
     and a column for each site in the alignment.
}
\description{
This function reads a fasta alignment file and converts it
     to an easy-to-read data frame object. Each individual 
     must have the same sequence length. The file cannot
     contain any comments.
}
\examples{
read.fasta.alignment()
}
\keyword{DNA}
\keyword{alignment}
\keyword{fasta}
\keyword{read}
