% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nexus.alignment.R
\name{read.nexus.alignment}
\alias{read.nexus.alignment}
\title{Read a Nexus alignment file}
\usage{
read.nexus.alignment(file)
}
\arguments{
\item{file}{The name of the file to be read. The file name 
should be provided within quotes, either as an absolute 
file path, or a relative file path given the current 
working directory, getwd().}
}
\value{
Returns a data frame with taxon names as row names 
     and a column for each site in the alignment.
}
\description{
This function reads a nexus alignment file and converts it
     to an easy-to-read data frame object.
}
\examples{
read.nexus.alignment()
}
\keyword{DNA}
\keyword{alignment}
\keyword{nexus}
\keyword{read}
